requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
psycopg2-binary>=2.9.0
PyJWT>=2.8.0
bcrypt>=4.0.0
pydantic>=2.0.0
